requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
pi==0.1.2
Pillow==10.0.1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
gunicorn==21.2.0
whitenoise==6.4.0
